transfer code from SampleTable to InvoiceItemTable, prep for PR
